Added test #7

+++ src/main/java/com/big_brother/models/UserSpied.java
@@ -38,7 +38,7 @@ public class UserSpied {
         return periodicity;
     }
 
-    public void setPeriodicity(int periodicity) {
+    public void setPeriodicity(long periodicity) {
         this.periodicity = periodicity;
     }
 }
